requirements.txt
beautifulsoup4
selenium
webdriver-manager
twilio
pandas
pyyaml
